# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM Pay API",
        "description": "API pour l'application OM Pay - Gestion des paiements et transferts",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        },
        {
            "url": "https://om-pay-mk4h.onrender.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/{numeroCompte}/transfert/initier": {
            "post": {
                "tags": [
                    "Transferts"
                ],
                "summary": "Initier un transfert",
                "description": "Crée une demande de transfert d'argent",
                "operationId": "02feb6468ab464f229a3b801da09d64f",
                "parameters": [
                    {
                        "name": "numeroCompte",
                        "in": "path",
                        "description": "Numéro de compte de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "7735434534"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numeroTelephoneDestinataire",
                                    "montant"
                                ],
                                "properties": {
                                    "numeroTelephoneDestinataire": {
                                        "type": "string",
                                        "example": "+221701234567"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    },
                                    "description": {
                                        "description": "Description optionnelle",
                                        "type": "string",
                                        "example": "Paiement loyer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfert initié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "idTransfert": {
                                                    "type": "string"
                                                },
                                                "montant": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "frais": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "statut": {
                                                    "type": "string",
                                                    "example": "EN_ATTENTE"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation ou solde insuffisant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/{numeroCompte}/transfert/{idTransfert}/confirmer": {
            "post": {
                "tags": [
                    "Transferts"
                ],
                "summary": "Confirmer un transfert",
                "description": "Confirme et exécute un transfert en attente",
                "operationId": "b6ab515041663555260a83cb8e8b4bc4",
                "parameters": [
                    {
                        "name": "numeroCompte",
                        "in": "path",
                        "description": "Numéro de compte de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "7735434534"
                        }
                    },
                    {
                        "name": "idTransfert",
                        "in": "path",
                        "description": "ID du transfert à confirmer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codePin"
                                ],
                                "properties": {
                                    "codePin": {
                                        "description": "Code PIN de l'utilisateur",
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfert confirmé et exécuté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "idTransfert": {
                                                    "type": "string"
                                                },
                                                "statut": {
                                                    "type": "string",
                                                    "example": "REUSSI"
                                                },
                                                "dateExecution": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "PIN incorrect ou transfert déjà traité",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/{numeroCompte}/transfert/{idTransfert}/annuler": {
            "delete": {
                "tags": [
                    "Transferts"
                ],
                "summary": "Annuler un transfert",
                "description": "Annule un transfert en attente",
                "operationId": "3ddc8f80eb351cf7ce532435c167f190",
                "parameters": [
                    {
                        "name": "numeroCompte",
                        "in": "path",
                        "description": "Numéro de compte de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "7735434534"
                        }
                    },
                    {
                        "name": "idTransfert",
                        "in": "path",
                        "description": "ID du transfert à annuler",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transfert annulé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "idTransfert": {
                                                    "type": "string"
                                                },
                                                "statut": {
                                                    "type": "string",
                                                    "example": "ANNULE"
                                                },
                                                "dateAnnulation": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Transfert ne peut pas être annulé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transfert déjà traité"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}